finish caloriesburned module

+++ ecofriend/lib/models/footprint.dart
@@ -1,12 +1,14 @@
 // To parse this JSON data, do
 //
-//     final footprint = footprintFromJson(jsonString);
+//     final Footprint = FootprintFromJson(jsonString);
 
 import 'dart:convert';
 
-List<Footprint> footprintFromJson(String str) => List<Footprint>.from(json.decode(str).map((x) => Footprint.fromJson(x)));
+List<Footprint> footprintFromJson(String str) =>
+    List<Footprint>.from(json.decode(str).map((x) => Footprint.fromJson(x)));
 
-String footprintToJson(List<Footprint> data) => json.encode(List<dynamic>.from(data.map((x) => x.toJson())));
+String footprintToJson(List<Footprint> data) =>
+    json.encode(List<dynamic>.from(data.map((x) => x.toJson())));
 
 class Footprint {
   Footprint({
@@ -20,16 +22,16 @@ class Footprint {
   Fields fields;
 
   factory Footprint.fromJson(Map<String, dynamic> json) => Footprint(
-    model: json["model"],
-    pk: json["pk"],
-    fields: Fields.fromJson(json["fields"]),
-  );
+        model: json["model"],
+        pk: json["pk"],
+        fields: Fields.fromJson(json["fields"]),
+      );
 
   Map<String, dynamic> toJson() => {
-    "model": model,
-    "pk": pk,
-    "fields": fields.toJson(),
-  };
+        "model": model,
+        "pk": pk,
+        "fields": fields.toJson(),
+      };
 }
 
 class Fields {
@@ -45,29 +47,29 @@ class Fields {
 
   int user;
   DateTime datetime;
-  int mileage;
-  int carbon;
+  double mileage;
+  double carbon;
   bool onFoot;
   String datetimeShow;
-  int toOrder;
+  double toOrder;
 
   factory Fields.fromJson(Map<String, dynamic> json) => Fields(
-    user: json["user"],
-    datetime: DateTime.parse(json["datetime"]),
-    mileage: json["mileage"],
-    carbon: json["carbon"],
-    onFoot: json["onFoot"],
-    datetimeShow: json["datetime_show"],
-    toOrder: json["to_order"],
-  );
+        user: json["user"],
+        datetime: DateTime.parse(json["datetime"]),
+        mileage: json["mileage"],
+        carbon: json["carbon"],
+        onFoot: json["onFoot"],
+        datetimeShow: json["datetime_show"],
+        toOrder: json["to_order"],
+      );
 
   Map<String, dynamic> toJson() => {
-    "user": user,
-    "datetime": datetime.toIso8601String(),
-    "mileage": mileage,
-    "carbon": carbon,
-    "onFoot": onFoot,
-    "datetime_show": datetimeShow,
-    "to_order": toOrder,
-  };
+        "user": user,
+        "datetime": datetime.toIso8601String(),
+        "mileage": mileage,
+        "carbon": carbon,
+        "onFoot": onFoot,
+        "datetime_show": datetimeShow,
+        "to_order": toOrder,
+      };
 }
